requirements.txt
streamlit-option-menu
scikit-learn
numpy
pandas
re
requests
joblib
beautifulsoup
tldextract
